# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Homestay App API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # PROPERTIES
...
  /bookings:
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (inquiry or instant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/status/{referenceNumber}:
    get:
      operationId: getBookingByReference
      tags: [bookings]
      summary: Track booking by reference number (no login required)
      parameters:
        - name: referenceNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/guest:
    get:
      operationId: getGuestBookings
      tags: [bookings]
      summary: Get bookings for the current guest
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Guest bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingDetail"
  /bookings/host:
    get:
      operationId: getHostBookings
      tags: [bookings]
      summary: Get all bookings for host's properties
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: propertyId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Host bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingDetail"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking details
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
  /bookings/{bookingId}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (host approve/decline)
      security:
        - bearerAuth: []
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingStatusRequest"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # DASHBOARD
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [host, guest]
      required:
        - name
        - email
        - password
        - role
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        role:
          type: string
          enum: [host, guest]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt

    # PROPERTY SCHEMAS
    Property:
      type: object
      properties:
        id:
          type: string
        hostId:
          type: string
        name:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        photos:
          type: array
          items:
            type: string
        description:
          type: string
        amenities:
          type: array
          items:
            type: string
        mealsIncluded:
          type: boolean
        locationLat:
          type: number
        locationLng:
          type: number
        nearbyAttractions:
          type: array
          items:
            type: string
        upiId:
          type: string
        bookingMode:
          type: string
          enum: [inquiry, instant]
        cancellationPolicy:
          type: string
        status:
          type: string
          enum: [active, inactive]
        averageRating:
          type: number
        reviewCount:
          type: integer
        minPrice:
          type: number
        createdAt:
          type: string
      required:
        - id
        - hostId
        - name
        - city
        - state
        - bookingMode
        - status
        - createdAt
...
    Room:
      type: object
      properties:
        id:
          type: string
        propertyId:
          type: string
        name:
          type: string
        type:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        photos:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [available, occupied, maintenance]
        createdAt:
          type: string
      required:
        - id
        - propertyId
        - name
        - type
        - pricePerNight
        - capacity
        - status
        - createdAt
...
    Booking:
      type: object
      properties:
        id:
          type: string
        roomId:
          type: string
        guestId:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        totalAmount:
          type: number
        paymentStatus:
          type: string
          enum: [pending, paid, refunded]
        specialRequests:
          type: string
        referenceNumber:
          type: string
        guestName:
          type: string
        guestEmail:
          type: string
        guestMobile:
          type: string
        guestCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - roomId
        - checkIn
        - checkOut
        - status
        - totalAmount
        - paymentStatus
        - referenceNumber
        - createdAt
    BookingDetail:
      allOf:
        - $ref: "#/components/schemas/Booking"
        - type: object
          properties:
            room:
              $ref: "#/components/schemas/Room"
            property:
              $ref: "#/components/schemas/Property"
            guest:
              $ref: "#/components/schemas/User"
    CreateBookingRequest:
      type: object
      properties:
        roomId:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        guestCount:
          type: integer
        guestName:
          type: string
        guestEmail:
          type: string
        guestMobile:
          type: string
        specialRequests:
          type: string
        guestId:
          type: string
      required:
        - roomId
        - checkIn
        - checkOut
        - guestCount
        - guestName
        - guestEmail
        - guestMobile
    UpdateBookingStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [confirmed, cancelled, completed]
      required:
        - status

    # HOST DASHBOARD
...